==> notebooks/03_final_answer.ipynb <==
from crewai import Agent, Task, Crew, Process, LLM
from Config import basic_llm
import os
from dotenv import load_dotenv

print(basic_llm)

user_input = "who is the president of Egypt"

# Load the Key token
_ = load_dotenv(override=True)
GEMINI_API_KEY = os.getenv('GEMINI_API_KEY')

basic_llm = LLM(model='gemini/gemini-2.0-flash', temperature=0)

final_answer_agent = Agent(
  role= "Final Answer Agent",
  goal= "To answer the user's question",
  backstory= "I am the final answer agent, I am responsible for answering the user's question",
  llm= basic_llm,
  )

final_answer_task = Task(
  description= "Answer the user's question: {user_input}",
  expected_output= "The answer to the user's question",
  agent= final_answer_agent,
)

ai_crew = Crew(
    agents=[
      final_answer_agent
    ],
    tasks=[
      final_answer_task
    ],
    process=Process.sequential,
)

result = ai_crew.kickoff(inputs={
  'user_input': user_input
})

result.raw
