==> notebooks/05_parse_html.ipynb <==
import json

with open('./ai-agent-output/sales_agent_output_json.json') as f:
    books = json.load(f)

with open('../template/one_card_template.txt') as f:
    one_cand_template = f.read()

with open('../template/cards_template.txt') as f:
    all_cand_template = f.read()

books

all_cards = ''
for book in books['books'][:3]:
  new_card_template = one_cand_template.replace("img_url" , book['Image_URL'])
  new_card_template = new_card_template.replace("book_title" , book['Book_title'])
  new_card_template = new_card_template.replace("book_author" , book['Book_author'])
  new_card_template = new_card_template.replace("discount_percentage" , str(book['discount_percentage']))
  new_card_template = new_card_template.replace("price_befor_diccount" , str(book['price']))
  new_card_template = new_card_template.replace("price_after_discount" , str(book['price_after_discount']))
  
  all_cards += new_card_template + '\n'

print(all_cards)

new_template = all_cand_template.replace('put_your_cards_here', all_cards)

print(new_template)

with open('../template/cards.html', 'w') as file:
  file.write(new_template)

test_dict = {
  "personalized_sales_assistance": "اهلا بك! كل شيء على ما يرام، شكرا لسؤالك. ما الذي تبحث عنه اليوم؟ هل لديك أي أنواع مفضلة أو مؤلفين معينين في ذهنك؟ أو ربما قرأت كتابًا جيدًا مؤخرًا وتود شيئًا مشابهًا؟ أخبرني وسأكون سعيدًا بتقديم بعض التوصيات.",
  "book_recommendations": {
    "books": []
  }
}

book_rec = test_dict.get('book_recommendations')
